clean up list array module
