+++ vortex-array/src/arrays/list/tests.rs
@@ -9,10 +9,14 @@ use vortex_dtype::PType::I32;
 use vortex_dtype::{DType, Nullability};
 use vortex_error::VortexUnwrap;
 use vortex_mask::Mask;
+use vortex_scalar::Scalar;
 
 use super::*;
+use crate::IntoArray;
 use crate::arrays::PrimitiveArray;
+use crate::builders::{ArrayBuilder, ListBuilder};
 use crate::compute::filter;
+use crate::validity::Validity;
 
 #[test]
 fn test_empty_list_array() {
