+++ vortex-array/src/arrays/list/vtable/array.rs
@@ -1,0 +1,22 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_dtype::DType;
+
+use crate::arrays::{ListArray, ListVTable};
+use crate::stats::StatsSetRef;
+use crate::vtable::ArrayVTable;
+
+impl ArrayVTable<ListVTable> for ListVTable {
+    fn len(array: &ListArray) -> usize {
+        array.offsets.len().saturating_sub(1)
+    }
+
+    fn dtype(array: &ListArray) -> &DType {
+        &array.dtype
+    }
+
+    fn stats(array: &ListArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+}
