+++ vortex-array/src/arrays/list/vtable/canonical.rs
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use crate::Canonical;
+use crate::arrays::{ListArray, ListVTable};
+use crate::vtable::CanonicalVTable;
+
+impl CanonicalVTable<ListVTable> for ListVTable {
+    fn canonicalize(array: &ListArray) -> Canonical {
+        Canonical::List(array.clone())
+    }
+}
