+++ vortex-array/src/arrays/list/vtable/mod.rs
@@ -1,0 +1,41 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use crate::arrays::ListArray;
+use crate::vtable::{NotSupported, VTable, ValidityVTableFromValidityHelper};
+use crate::{EncodingId, EncodingRef, vtable};
+
+mod array;
+mod canonical;
+mod operations;
+mod serde;
+mod validity;
+mod visitor;
+
+vtable!(List);
+
+impl VTable for ListVTable {
+    type Array = ListArray;
+    type Encoding = ListEncoding;
+
+    type ArrayVTable = Self;
+    type CanonicalVTable = Self;
+    type OperationsVTable = Self;
+    type ValidityVTable = ValidityVTableFromValidityHelper;
+    type VisitorVTable = Self;
+    type ComputeVTable = NotSupported;
+    type EncodeVTable = NotSupported;
+    type PipelineVTable = NotSupported;
+    type SerdeVTable = Self;
+
+    fn id(_encoding: &Self::Encoding) -> EncodingId {
+        EncodingId::new_ref("vortex.list")
+    }
+
+    fn encoding(_array: &Self::Array) -> EncodingRef {
+        EncodingRef::new_ref(ListEncoding.as_ref())
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct ListEncoding;
